+++ src/fns/index.ts
@@ -3,5 +3,6 @@ export * from './array.fns';
 export * from './number.fns';
 export * from './date.fns';
 export * from './color.fns';
+export * from './browser.fns';
 export * from './utility.fns';
 export * from './detect-incognito.fns';